Refactor BackFlow Pfaffian functions into dedicated source
