+++ src/mVMC/makefile_src
@@ -23,6 +23,7 @@ OBJS = \
 
 SOURCES = \
 backflow.c \
+backflow_pf.c \
 backflow_slater.c \
 average.c \
 avevar.c \
