+++ src/mVMC/matrix.c
@@ -387,129 +387,6 @@ int calculateMAll_child_fcmp(const int *eleIdx, const int qpStart, const int qpE
   return info;
 }
 
-int CalculateMAll_BF_fcmp(const int *eleIdx, const int qpStart, const int qpEnd) {
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-
-  int info = 0;
-
-  double complex*myBufM;
-  double complex*myWork;
-  int *myIWork;
-  int myInfo;
-  double *myRWork;
-
-  RequestWorkSpaceThreadInt(Nsize);
-  RequestWorkSpaceThreadComplex(Nsize*Nsize+LapackLWork);
-
-  RequestWorkSpaceThreadDouble(LapackLWork); // TBC for rwork
-
-#pragma omp parallel default(shared)              \
-  private(myIWork,myWork,myRWork, myInfo,myBufM) //TODO: Check to add myRWork is correct or not.
-  {
-    myIWork = GetWorkSpaceThreadInt(Nsize);
-    myBufM = GetWorkSpaceThreadComplex(Nsize*Nsize);
-    myWork = GetWorkSpaceThreadComplex(LapackLWork);
-
-    myRWork = GetWorkSpaceThreadDouble(LapackLWork); //TBC for rwork
-
-#pragma omp for private(qpidx)
-    for(qpidx=0;qpidx<qpNum;qpidx++) {
-      if(info!=0) continue;
-
-      myInfo = calculateMAll_BF_fcmp_child(eleIdx, qpStart, qpEnd, qpidx,
-          myBufM, myIWork, myWork, LapackLWork, myRWork, PfM, InvM);
-      if(myInfo!=0) {
-#pragma omp critical
-        info=myInfo;
-      }
-    }
-  }
-
-  ReleaseWorkSpaceThreadInt();
-  ReleaseWorkSpaceThreadComplex();
-  ReleaseWorkSpaceThreadDouble();
-  return info;
-}
-
-int calculateMAll_BF_fcmp_child(
-       const int *eleIdx,
-       const int qpStart,
-       const int qpEnd,
-       const int qpidx,
-       double complex*bufM,
-       int *iwork,
-       double complex*work,
-       int lwork,
-       double* rwork,
-       double complex* PfM,
-       double complex* InvM
-       )
-{
-#pragma procedure serial
-  /* const int qpNum = qpEnd-qpStart; */
-  int msi,msj;
-  int rsi,rsj;
-
-  char uplo='U', mthd='P';
-  int m,n,lda,nsq,one,info=0;
-  double complex pfaff,minus_one;
-
-  /* optimization for Kei */
-  const int nsize = Nsize;
-
-  const double complex *sltE = SlaterElmBF + (qpidx+qpStart)*Nsite2*Nsite2;
-  const double complex *sltE_i;
-
-  double complex*invM = InvM + qpidx*Nsize*Nsize;
-  double complex*invM_i;
-
-  double complex*bufM_i, *bufM_i2;
-
-  m=n=lda=Nsize;
-  nsq=n*n;
-  one=1;
-  minus_one=-1.0;
-
-  /* store bufM */
-  /* Note that bufM is column-major and skew-symmetric. */
-  /* bufM[msj][msi] = -sltE[rsi][rsj] */
-#pragma loop noalias
-  for(msi=0;msi<nsize;msi++) {
-    rsi = eleIdx[msi] + (msi/Ne)*Nsite;
-    bufM_i = bufM + msi*Nsize;
-    sltE_i = sltE + rsi*Nsite2;
-#pragma loop norecurrence
-    for(msj=0;msj<nsize;msj++) {
-      rsj = eleIdx[msj] + (msj/Ne)*Nsite;
-      bufM_i[msj] = -sltE_i[rsj];
-    }
-  }
-
-  /* Pfaffian/inverse computed separately. */
-  /* Copy bufM to invM before using bufM to compute Pfaffian. */
-  for(msi=0;msi<nsize*nsize;msi++)
-    invM[msi] = bufM[msi];
-
-  /* Calculate Pf M */
-  M_ZSKPFA(&uplo, &mthd, &n, bufM, &lda, &pfaff, iwork, work, &lwork, rwork, &info);
-
-  if(info!=0) return info;
-  if(!(isfinite(creal(pfaff)) && isfinite(cimag(pfaff)))) return qpidx+1;
-  PfM[qpidx] = pfaff;
-
-  /* Calculate inverse. */
-  M_ZGETRF(&m, &n, invM, &lda, iwork, &info); /* ipiv = iwork */
-  if(info!=0) return info;
-  M_ZGETRI(&n, invM, &lda, iwork, work, &lwork, &info);
-  if(info!=0) return info;
-
-  /* InvM -> InvM(T) -> -InvM */
-  M_ZSCAL(&nsq, &minus_one, invM, &one);
-
-  return info;
-}
-
 int CalculateMAll_BF_fsz(const int *eleIdx, const int *eleSpn, const int qpStart, const int qpEnd) {
   const int qpNum = qpEnd-qpStart;
   int qpidx;
@@ -976,113 +853,5 @@ int calculateMAll_child_real(const int *eleIdx, const int qpStart, const int qpE
   return info;
 }
 
-int CalculateMAll_BF_real(const int *eleIdx, const int qpStart, const int qpEnd){
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-
-  int info = 0;
-
-  double *myBufM;
-  double *myWork;
-  int *myIWork;
-  int myInfo;
-
-  RequestWorkSpaceThreadInt(Nsize);
-  RequestWorkSpaceThreadDouble(Nsize*Nsize+LapackLWork);
-
-  /* Keep real BackFlow MultiQP rebuilds serialized; Linux CI shows the
-     concurrent QP rebuild path can corrupt the identity-energy run. */
-#pragma omp parallel if(qpNum == 1) default(shared)              \
-  private(myIWork,myWork,myInfo,myBufM)
-  {
-    myIWork = GetWorkSpaceThreadInt(Nsize);
-    myBufM = GetWorkSpaceThreadDouble(Nsize*Nsize);
-    myWork = GetWorkSpaceThreadDouble(LapackLWork);
-
-#pragma omp for private(qpidx)
-    for(qpidx=0;qpidx<qpNum;qpidx++) {
-      if(info!=0) continue;
-
-      myInfo = calculateMAll_BF_real_child(eleIdx, qpStart, qpEnd, qpidx,
-          myBufM, myIWork, myWork, LapackLWork, PfM_real, InvM_real);
-      if(myInfo!=0) {
-#pragma omp critical
-        info=myInfo;
-      }
-    }
-  }
-
-  ReleaseWorkSpaceThreadInt();
-  ReleaseWorkSpaceThreadDouble();
-  return info;
-}
-
-int calculateMAll_BF_real_child(const int *eleIdx, const int qpStart, const int qpEnd, const int qpidx,
-    double *bufM, int *iwork, double *work, int lwork, double* PfM_real, double* InvM_real) {
-#pragma procedure serial
-  /* const int qpNum = qpEnd-qpStart; */
-  int msi,msj;
-  int rsi,rsj;
-
-  char uplo='U', mthd='P';
-  int m,n,nsq,one,lda,info=0;
-  double pfaff,minus_one;
-
-  /* optimization for Kei */
-  const int nsize = Nsize;
-
-  const double *sltE = SlaterElmBF_real + (qpidx+qpStart)*Nsite2*Nsite2;
-  const double *sltE_i;
-
-  double *invM = InvM_real + qpidx*Nsize*Nsize;
-  double *invM_i;
-
-  double *bufM_i, *bufM_i2;
-
-  m=n=lda=Nsize;
-  nsq=n*n;
-  one=1;
-  minus_one=-1.0;
-
-  /* store bufM */
-  /* Note that bufM is column-major and skew-symmetric. */
-  /* bufM[msj][msi] = -sltE[rsi][rsj] */
-#pragma loop noalias
-  for(msi=0;msi<nsize;msi++) {
-    rsi = eleIdx[msi] + (msi/Ne)*Nsite;
-    bufM_i = bufM + msi*Nsize;
-    sltE_i = sltE + rsi*Nsite2;
-#pragma loop norecurrence
-    for(msj=0;msj<nsize;msj++) {
-      rsj = eleIdx[msj] + (msj/Ne)*Nsite;
-      bufM_i[msj] = -sltE_i[rsj];
-    }
-  }
-
-  /* Pfaffian/inverse computed separately. */
-  /* Copy bufM to invM before using bufM to compute Pfaffian. */
-  for(msi=0;msi<nsize*nsize;msi++)
-    invM[msi] = bufM[msi];
-
-  /* Calculate Pf M */
-  M_DSKPFA(&uplo, &mthd, &n, bufM, &lda, &pfaff, iwork, work, &lwork, &info);
-
-  if(info!=0) return info;
-  if(!isfinite(pfaff)) return qpidx+1;
-  PfM_real[qpidx] = pfaff;
-
-  /* Compute inverse. */
-  M_DGETRF(&m, &n, invM, &lda, iwork, &info); /* ipiv = iwork */
-  if(info!=0) return info;
-  M_DGETRI(&n, invM, &lda, iwork, work, &lwork, &info);
-  if(info!=0) return info;
-
-  // InvM -> InvM' = -InvM
-  M_DSCAL(&nsq, &minus_one, invM, &one);
-
-  return info;
-}
-
-//==============e real =============//
 
 #endif
