+++ src/mVMC/pfupdate.c
@@ -29,13 +29,6 @@ along with this program. If not, see http://www.gnu.org/licenses/.
 #ifndef _PFUDATE_SRC
 #define _PFUDATE_SRC
 
-double complex updateMAll_BF_fcmp_child(
-        const int qpidx,
-        const int globalQpidx,
-        const int n, const int *msa,
-        const int *eleIdx);
-
-
 /* Calculate new pfaffian. The ma-th electron with spin s hops. */
 void CalculateNewPfM(const int ma, const int s, double complex *pfMNew, const int *eleIdx,
                      const int qpStart, const int qpEnd) {
@@ -228,402 +221,5 @@ void updateMAll_child(const int ma, const int s, const int *eleIdx,
 
 /* Calculate new pfaffian with Backflow effects.
    The ma-th electron with spin s hops from ra to rb */
-void CalculateNewPfMBFWithStride(const int *icount, const int *msaTmp, const int msaStride,
-                       double complex* pfMNew, const int *eleIdx,
-                       const int qpStart, const int qpEnd, const double complex* bufM) {
-  //#pragma procedure serial
-  int i;
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-  int globalQpidx;
-  int *msa;
-
-  for(qpidx=0;qpidx<qpNum;qpidx++) {
-    globalQpidx = qpidx + qpStart;
-    //Store msa//
-    msa=(int *)malloc(sizeof(int)*icount[globalQpidx]);
-    //printf("Total=%d\n",icount[qpidx]);
-    for(i=0;i<icount[globalQpidx];i++){
-      msa[i] = msaTmp[i+globalQpidx*msaStride];
-      //printf("hop[%d]=%d\n",i,msa[i]);
-    }
-
-    /* calculateNewPfM */
-    pfMNew[qpidx] = calculateNewPfMBFN4_child(qpidx,globalQpidx,icount[globalQpidx],msa,eleIdx,bufM);
-
-    free(msa);
-  }
-
-  return;
-}
-
-void CalculateNewPfMBF(const int *icount, const int *msaTmp,
-                       double complex* pfMNew, const int *eleIdx,
-                       const int qpStart, const int qpEnd, const double complex* bufM) {
-  CalculateNewPfMBFWithStride(icount, msaTmp, Nsize, pfMNew, eleIdx, qpStart, qpEnd, bufM);
-}
-
-double complex calculateNewPfMBFN4_child(const int qpidx, const int globalQpidx, const int n, const int *msa,
-                                         const int *eleIdx, const double complex* bufM)
-{
-  const int nsize = Nsize;
-  const int n2 = 2*n;
-  const double complex *sltE;
-  const double complex *sltE_k;
-  double complex *invM;
-  double complex *invM_i, *invM_k, *invM_l;
-
-  double complex *vec; /* vec[n][nsize] */
-  //double complex vec[n*nsize]; /* vec[n][nsize] */
-  double complex *vec_k, *vec_l;
-  //double complex mat[n2*n2]; /* mat[n2][n2] */
-  double complex *mat; /* mat[n2][n2] */
-  double complex *mat_k;
-  //double complex matUV[n2*nsize]; /* mat[n2][nsize] */
-  double sgn;
-
-  int rsi,rsk,msi,msj,k,l;
-  double complex val,tmp;
-
-  /* for ZSKPFA */
-  char uplo='U', mthd='P';
-  int nn,lda,info=0;
-  double complex pfaff;
-  int iwork[n2];
-  //double complex work[n2*n2]; /* [n2][n2] */
-  double complex *work; /* [n2][n2] */
-  double rwork[n2*n2]; /* [n2][n2] */
-  int lwork = n2*n2;
-  nn=lda=n2;
-
-  sltE = bufM + globalQpidx*Nsite2*Nsite2;
-  invM = InvM + qpidx*Nsize*Nsize;
-
-  vec = (double complex*)malloc(sizeof(double complex)*n*nsize);
-  //invMat = (double *)malloc(sizeof(double)*n2*n2);
-  mat = (double complex*)malloc(sizeof(double complex)*n2*n2);
-  //matUV = (double *)malloc(sizeof(double)*n2*nsize);
-  work = (double complex*)malloc(sizeof(double complex)*n2*n2);
-
-  //#pragma loop noalias
-  for(k=0;k<n;k++) {
-    rsk = eleIdx[msa[k]] + (msa[k]/Ne)*Nsite;
-    sltE_k = sltE + rsk*Nsite2;
-    vec_k = vec + k*nsize;
-    //#pragma loop norecurrence
-    for(msi=0;msi<nsize;msi++) {
-      rsi = eleIdx[msi] + (msi/Ne)*Nsite;
-      vec_k[msi] = sltE_k[rsi];
-    }
-  }
-
-  /* X_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k;
-    vec_k = vec + k*nsize;
-    for(l=k+1;l<n;l++) {
-      vec_l = vec + l*nsize;
-      val = 0.0;
-      for(msi=0;msi<nsize;msi++) {
-        invM_i = invM + msi*nsize;
-        tmp = 0.0;
-        for(msj=0;msj<nsize;msj++) {
-          tmp += invM_i[msj] * vec_l[msj];
-        }
-        val += tmp * vec_k[msi];
-      }
-      mat_k[l] = val + vec_k[msa[l]];
-    }
-  }
-
-  /* Y_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k + n;
-    vec_k = vec + k*nsize;
-    for(l=0;l<n;l++) {
-      invM_l = invM + msa[l]*nsize;
-      val = 0.0;
-      for(msi=0;msi<nsize;msi++) {
-        val += vec_k[msi] * invM_l[msi];
-      }
-      mat_k[l] = val;
-    }
-  }
-
-  /* Z_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*(k+n) + n;
-    invM_k = invM + msa[k]*nsize;
-    for(l=k+1;l<n;l++) {
-      mat_k[l] = invM_k[msa[l]];
-    }
-  }
-
-  //#pragma loop noalias
-  for(k=0;k<n2;k++) {
-    //#pragma loop norecurrence
-    for(l=0;l<k;l++) {
-      mat[n2*k + l] = -mat[n2*l + k]; /* transpose */
-    }
-    mat[n2*k + k] = 0.0; /* diagonal elements */
-  }
-
-  /* Calculate Pf M */
-  //TODO: CHECK rwork is real <-- [R-Xu] Not needed anymore?
-  M_ZSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, rwork, &info);
-  //M_DSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, &info);
-
-  sgn = ( (n*(n-1)/2)%2==0 ) ? 1.0 : -1.0;
-
-  //free(matUV);
-  free(mat);
-  free(vec);
-  //free(invMat);
-  free(work);
-  return sgn * pfaff * PfM[qpidx];
-}
-
-void UpdateMAll_BF_fcmp(const int *icount, const int *msaTmp,
-                        double complex* pfMNew, const int *eleIdx,
-                        const int qpStart, const int qpEnd)
-{
-#pragma procedure serial
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-  int globalQpidx;
-  //double complex *sltE;
-  //double complex *sltE_i;
-  int *msa;
-  int i;
-  //int *hop;
-  //double complex diff;
-
-  for(qpidx=0;qpidx<qpNum;qpidx++) {
-    globalQpidx = qpidx + qpStart;
-    //Store msa//
-    msa=(int *)malloc(sizeof(int)*icount[globalQpidx]);
-    for(i=0;i<icount[globalQpidx];i++){
-      msa[i] = msaTmp[i+globalQpidx*Nsize];
-    }
-
-    /* calculateNewPfM */
-    pfMNew[qpidx] = updateMAll_BF_fcmp_child(qpidx,globalQpidx,icount[globalQpidx],msa,eleIdx);
-
-    free(msa);
-  }
-
-  return;
-}
-
-/* msa[k]-th electron hops from rsa[k] to eleIdx[msa[k]] */
-/* buffer size = n*Nsize */
-//void updateMAllBF3_child(const int qpidx, const int n, const int *msa,
-double complex updateMAll_BF_fcmp_child(
-        const int qpidx,
-        const int globalQpidx,
-        const int n, const int *msa,
-        const int *eleIdx)
-{
-  const int nsize = Nsize;
-  const int n2 = 2*n;
-  const double complex*sltE;
-  const double complex*sltE_k;
-  double complex*invM;
-  double complex *invM_i, *invM_k, *invM_l;
-
-  double complex*vec; /* vec[n][nsize] */
-  double complex*vec_k, *vec_l;
-  //double complex mat[n2*n2]; /* mat[n2][n2] */
-  double complex*mat; /* mat[n2][n2] */
-  double complex*mat_k;
-  //double complex invMat[n2*n2]; /* mat[n2][n2] */
-  double complex*invMat; /* mat[n2][n2] */
-  //double complex matUV[n2*nsize]; /* mat[n2][nsize] */
-  double complex*matUV; /* mat[n2][nsize] */
-  double complex*invMat_k, *matUV_i, *matUV_j;
-  double complex sgn;
-
-  int rsi,rsk,msi,msj,k,l;
-  double complex val,tmp;
-
-  /* for DSKPFA */
-  char uplo='U', mthd='P';
-  int m,nn,lda,info=0;
-  double complex pfaff;
-  int iwork[n2];
-  int iwork2[n2];
-  //double complex work[n2*n2]; /* [n2][n2] */
-  double rwork[n2*n2]; /* [n2][n2] */
-  double complex*work; /* [n2][n2] */
-  double complex*work2; /* [n2][n2] */
-  int lwork = n2*n2;
-  int lwork2 = n2;
-  m=nn=lda=n2;
-
-  sltE = SlaterElmBF + globalQpidx*Nsite2*Nsite2;
-  invM = InvM + qpidx*Nsize*Nsize;
-
-  //vec = bufferc; /* n*nsize */
-  vec = (double complex*)malloc(sizeof(double complex)*n*nsize);
-  invMat = (double complex*)malloc(sizeof(double complex)*n2*n2);
-  mat = (double complex*)malloc(sizeof(double complex)*n2*n2);
-  matUV = (double complex*)malloc(sizeof(double complex)*n2*nsize);
-  work = (double complex*)malloc(sizeof(double complex)*n2*n2);
-  work2 = (double complex*)malloc(sizeof(double complex)*n2);
-
-#pragma loop noalias
-  for(k=0;k<n;k++) {
-    rsk = eleIdx[msa[k]] + (msa[k]/Ne)*Nsite;
-    sltE_k = sltE + rsk*Nsite2;
-    vec_k = vec + k*nsize;
-#pragma loop norecurrence
-    for(msi=0;msi<nsize;msi++) {
-      rsi = eleIdx[msi] + (msi/Ne)*Nsite;
-      vec_k[msi] = sltE_k[rsi];
-    }
-  }
-
-  /* X_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k;
-    vec_k = vec + k*nsize;
-    for(l=k+1;l<n;l++) {
-      vec_l = vec + l*nsize;
-      val = 0.0;
-      for(msi=0;msi<nsize;msi++) {
-        invM_i = invM + msi*nsize;
-        tmp = 0.0;
-        for(msj=0;msj<nsize;msj++) {
-          tmp += invM_i[msj] * vec_l[msj];
-        }
-        val += tmp * vec_k[msi];
-      }
-      mat_k[l] = val + vec_k[msa[l]];
-    }
-  }
-
-  /* Y_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k + n;
-    vec_k = vec + k*nsize;
-    for(l=0;l<n;l++) {
-      invM_l = invM + msa[l]*nsize;
-      val = 0.0;
-      for(msi=0;msi<nsize;msi++) {
-        val += vec_k[msi] * invM_l[msi];
-      }
-      mat_k[l] = val;
-    }
-  }
-
-  /* Z_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*(k+n) + n;
-    invM_k = invM + msa[k]*nsize;
-    for(l=k+1;l<n;l++) {
-      mat_k[l] = invM_k[msa[l]];
-    }
-  }
-
-#pragma loop noalias
-  for(k=0;k<n2;k++) {
-#pragma loop norecurrence
-    for(l=0;l<k;l++) {
-      mat[n2*k + l] = -mat[n2*l + k]; /* transpose */
-    }
-    mat[n2*k + k] = 0.0; /* diagonal elements */
-  }
-
-  //#pragma loop noalias
-  for(k=0;k<n2;k++) {
-    //#pragma loop norecurrence
-    for(l=0;l<n2;l++) {
-      invMat[n2*k + l] = mat[n2*l + k];
-    }
-  }
-
-  /* matUV_ik */
-  for(msi=0;msi<nsize;msi++) {
-    matUV_i = matUV + n2*msi;
-    invM_i = invM + msi*nsize;
-    for(k=0;k<n;k++) {
-      vec_k = vec + k*nsize;
-      val = 0.0;
-      for(msj=0;msj<nsize;msj++) {
-        val -= invM_i[msj] * vec_k[msj];
-      }
-      if(msa[k]==msi){val -= 1.0;}
-      matUV_i[k] = val;
-      matUV_i[k+n] = invM_i[msa[k]];
-    }
-  }
-  /* DInv */
-  m=nn=lda=n2;
-  M_ZGETRF(&m, &nn, invMat, &lda, iwork2, &info); /* ipiv = iwork */
-  //M_DGETRF(&m, &nn, invMat, &lda, iwork2, &info); /* ipiv = iwork */
-  if(info!=0){
-    printf("Error M_ZGETRF %d\n",info);
-    //return;
-  }
-
-  M_ZGETRI(&nn, invMat, &lda, iwork2, work2, &lwork2, &info);
-  //M_DGETRI(&nn, invMat, &lda, iwork2, work2, &lwork2, &info);
-  if(info!=0){
-    printf("Error M_ZGETRI %d\n",info);
-    //return;
-  }
-
-  for(msi=0;msi<nsize;msi++) {
-    matUV_i = matUV + n2*msi;
-    invM_i = invM + msi*nsize;
-    for(msj=0;msj<nsize;msj++) {
-      matUV_j = matUV + n2*msj;
-      val = 0.0;
-      for(k=0;k<n2;k++) {
-        invMat_k = invMat + k*n2;
-        tmp=0.0;
-        for(l=0;l<n2;l++) {
-          tmp += invMat_k[l]*matUV_j[l];
-        }
-        val += matUV_i[k]*tmp;
-      }
-      invM_i[msj] -= val;
-    }
-  }
-
-  for(msi=0;msi<Ne;msi++) {
-    //invM_i = invM + msi*nsize;
-    for(msj=Ne;msj<Nsize;msj++) {
-      invM[msj*nsize+msi] = -invM[msi*nsize+msj];
-    }
-  }
-  for(msi=Ne;msi<Nsize;msi++) {
-    for(msj=Ne;msj<Nsize;msj++) {
-      invM[msi*nsize+msj] = 0.0;
-    }
-  }
-  for(msi=0;msi<Ne;msi++) {
-    for(msj=0;msj<Ne;msj++) {
-      invM[msi*nsize+msj] = 0.0;
-    }
-  }
-
-  /* Calculate Pf M */
-  M_ZSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, rwork, &info);
-  //M_DSKPFA(&uplo, &mthd, &nn, invM, &lda, &pfaff, iwork, work, &lwork, &info);
-  //M_DSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, &info);
-
-  sgn = ( (n*(n-1)/2)%2==0 ) ? 1.0 : -1.0;
-
-  free(matUV);
-  free(mat);
-  free(vec);
-  free(invMat);
-  free(work);
-  free(work2);
-
-  return sgn * pfaff * PfM[qpidx];
-}
-
 
 #endif
