+++ src/mVMC/pfupdate_real.c
@@ -36,33 +36,6 @@ void updateMAll_child_real(const int ma, const int s, const int *eleIdx,
                       const int qpStart, const int qpEnd, const int qpidx,
                       double *vec1, double *vec2);
 
-double calculateNewPfMBFN4_real_child(const int qpidx, const int globalQpidx, const int n, const int *msa,
-                                      const int *eleIdx, const double *bufM);
-double calculateNewPfMBFN4_real_child_vec(const int qpidx, const int n, const int *msa, const double *vec);
-static double calculateNewPfMBFN4_real_child_vec_w(const int qpidx, const int n, const int *msa,
-                                                   const double *vec, const double *w);
-
-double updateMAll_BF_real_child(const int qpidx, const int globalQpidx, const int n, const int *msa,
-                                const int *eleIdx);
-
-static void DgemmRowMajorNN(const int m, const int n, const int k,
-                            const double *a, const double *b,
-                            const double alpha, const double beta, double *c) {
-  char trans = 'N';
-  int fm = n, fn = m, fk = k;
-  int lda = n, ldb = k, ldc = n;
-  M_DGEMM(&trans, &trans, &fm, &fn, &fk, &alpha, b, &lda, a, &ldb, &beta, c, &ldc);
-}
-
-static void DgemmRowMajorNT(const int m, const int n, const int k,
-                            const double *a, const double *b,
-                            const double alpha, const double beta, double *c) {
-  char transa = 'T', transb = 'N';
-  int fm = n, fn = m, fk = k;
-  int lda = k, ldb = k, ldc = n;
-  M_DGEMM(&transa, &transb, &fm, &fn, &fk, &alpha, b, &lda, a, &ldb, &beta, c, &ldc);
-}
-
 /* Calculate new pfaffian. The ma-th electron with spin s hops. */
 void CalculateNewPfM_real(const int ma, const int s, double *pfMNew_real, const int *eleIdx,
                      const int qpStart, const int qpEnd) {
@@ -253,474 +226,3 @@ void updateMAll_child_real(const int ma, const int s, const int *eleIdx,
 
 /* Calculate new pfaffian with Backflow effects.
    The ma-th electron with spin s hops from ra to rb */
-void CalculateNewPfMBFWithStride_real(const int *icount, const int *msaTmp, const int msaStride,
-                       double *pfMNew, const int *eleIdx,
-                       const int qpStart, const int qpEnd, const double *bufM) {
-  //#pragma procedure serial
-  int i;
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-  int globalQpidx;
-  //double *sltE;
-  //double *sltE_i;
-  int *msa;
-  //int msi,msj,rsi,rsj,i;
-  //int *hop;
-  //double complex diff;
-
-  for(qpidx=0;qpidx<qpNum;qpidx++) {
-    globalQpidx = qpidx + qpStart;
-    //Store msa//
-    msa=(int *)malloc(sizeof(int)*icount[globalQpidx]);
-    //printf("Total=%d\n",icount[qpidx]);
-    for(i=0;i<icount[globalQpidx];i++){
-      msa[i] = msaTmp[i+globalQpidx*msaStride];
-      //printf("hop[%d]=%d\n",i,msa[i]);
-    }
-
-    /* calculateNewPfM */
-    pfMNew[qpidx] = calculateNewPfMBFN4_real_child(qpidx,globalQpidx,icount[globalQpidx],msa,eleIdx,bufM);
-
-    free(msa);
-  }
-  //icount = UpdateSlaterElmBFTmp3(ma, rb, ra, s, eleCfg, eleNum, msaTmp);
-
-  return;
-}
-
-void CalculateNewPfMBF_real(const int *icount, const int *msaTmp,
-                       double *pfMNew, const int *eleIdx,
-                       const int qpStart, const int qpEnd, const double *bufM) {
-  CalculateNewPfMBFWithStride_real(icount, msaTmp, Nsize, pfMNew, eleIdx, qpStart, qpEnd, bufM);
-}
-
-void CalculateNewPfMBFVecWithStride_real(const int *icount, const int *msaTmp, const int msaStride,
-                       double *pfMNew, const int qpStart, const int qpEnd,
-                       const double *vecM, const int vecStride) {
-  int i;
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-  int globalQpidx;
-  int *msa;
-  const double *vec;
-
-  for(qpidx=0;qpidx<qpNum;qpidx++) {
-    globalQpidx = qpidx + qpStart;
-    msa=(int *)malloc(sizeof(int)*icount[globalQpidx]);
-    for(i=0;i<icount[globalQpidx];i++){
-      msa[i] = msaTmp[i+globalQpidx*msaStride];
-    }
-
-    vec = vecM + globalQpidx*vecStride*Nsize;
-    pfMNew[qpidx] = calculateNewPfMBFN4_real_child_vec(qpidx,icount[globalQpidx],msa,vec);
-
-    free(msa);
-  }
-
-  return;
-}
-
-void CalculateNewPfMBFVec_real(const int *icount, const int *msaTmp,
-                       double *pfMNew, const int qpStart, const int qpEnd,
-                       const double *vecM) {
-  CalculateNewPfMBFVecWithStride_real(icount, msaTmp, Nsize, pfMNew, qpStart, qpEnd, vecM, Nsize);
-}
-
-void CalculateNewPfMBFVecBatched_real(const int batchSize, const int *icount, const int *msaTmp,
-                       double *pfMNew, const int qpStart, const int qpEnd,
-                       const double *vecM, double *vecStack, double *wStack) {
-  const int nsize = Nsize;
-  const int qpNum = qpEnd-qpStart;
-  int qpidx, globalQpidx, batchIdx, row, totalRows, offset;
-  int rowOffset[batchSize];
-  int n;
-  const int *msa;
-  const double *termVec;
-  const double *invM;
-
-  for(qpidx=0;qpidx<qpNum;qpidx++) {
-    globalQpidx = qpidx + qpStart;
-    totalRows = 0;
-    for(batchIdx=0;batchIdx<batchSize;batchIdx++) {
-      n = icount[batchIdx*NQPFull + globalQpidx];
-      rowOffset[batchIdx] = totalRows;
-      termVec = vecM + ((size_t)batchIdx*NQPFull + globalQpidx)*nsize*nsize;
-      for(row=0;row<n;row++) {
-        memcpy(vecStack + (totalRows+row)*nsize, termVec + row*nsize, sizeof(double)*nsize);
-      }
-      totalRows += n;
-    }
-
-    if(totalRows > 0) {
-      invM = InvM_real + qpidx*nsize*nsize;
-      DgemmRowMajorNT(totalRows, nsize, nsize, vecStack, invM, 1.0, 0.0, wStack);
-    }
-
-    for(batchIdx=0;batchIdx<batchSize;batchIdx++) {
-      n = icount[batchIdx*NQPFull + globalQpidx];
-      offset = rowOffset[batchIdx];
-      msa = msaTmp + ((size_t)batchIdx*NQPFull + globalQpidx)*nsize;
-      pfMNew[batchIdx*NQPFull + globalQpidx] =
-        calculateNewPfMBFN4_real_child_vec_w(qpidx, n, msa,
-                                             vecStack + offset*nsize,
-                                             wStack + offset*nsize);
-    }
-  }
-
-  return;
-}
-
-/* msa[k]-th electron hops from rsa[k] to eleIdx[msa[k]] */
-/* buffer size = n*Nsize */
-//double complex calculateNewPfMBFN_child(const int qpidx, const int n, const int *msa,
-//                              const int *eleIdx, double *rwork, double complex *bufferc) {
-double calculateNewPfMBFN4_real_child(const int qpidx, const int globalQpidx, const int n, const int *msa,
-                                      const int *eleIdx, const double *bufM) {
-  const int nsize = Nsize;
-  const double *sltE;
-  const double *sltE_k;
-
-  double *vec; /* vec[n][nsize] */
-  double *vec_k;
-
-  int rsi,rsk,msi,k;
-  double pfnew;
-
-  sltE = bufM + globalQpidx*Nsite2*Nsite2;
-
-  vec = (double *)malloc(sizeof(double)*n*nsize);
-
-  //#pragma loop noalias
-  for(k=0;k<n;k++) {
-    rsk = eleIdx[msa[k]] + (msa[k]/Ne)*Nsite;
-    sltE_k = sltE + rsk*Nsite2;
-    vec_k = vec + k*nsize;
-    //#pragma loop norecurrence
-    for(msi=0;msi<nsize;msi++) {
-      rsi = eleIdx[msi] + (msi/Ne)*Nsite;
-      vec_k[msi] = sltE_k[rsi];
-    }
-  }
-
-  pfnew = calculateNewPfMBFN4_real_child_vec(qpidx,n,msa,vec);
-
-  free(vec);
-  return pfnew;
-}
-
-double calculateNewPfMBFN4_real_child_vec(const int qpidx, const int n, const int *msa, const double *vec) {
-  const int nsize = Nsize;
-  double *invM;
-  double *w; /* w[n][nsize] = vec * invM^T */
-  double pfnew;
-
-  invM = InvM_real + qpidx*Nsize*Nsize;
-
-  w = (double *)malloc(sizeof(double)*n*nsize);
-
-  DgemmRowMajorNT(n, nsize, nsize, vec, invM, 1.0, 0.0, w);
-
-  pfnew = calculateNewPfMBFN4_real_child_vec_w(qpidx,n,msa,vec,w);
-
-  free(w);
-  return pfnew;
-}
-
-static double calculateNewPfMBFN4_real_child_vec_w(const int qpidx, const int n, const int *msa,
-                                                   const double *vec, const double *w) {
-  const int nsize = Nsize;
-  const int n2 = 2*n;
-  double *invM;
-  double *invM_k;
-
-  const double *vec_k;
-  double *mat; /* mat[n2][n2] */
-  double *mat_k;
-  double sgn;
-
-  int msi,k,l;
-  double val;
-
-  /* for ZSKPFA */
-  char uplo='U', mthd='P';
-  int nn,lda,info=0;
-  double pfaff;
-  int iwork[n2];
-  double *work; /* [n2][n2] */
-  int lwork = n2*n2;
-  nn=lda=n2;
-
-  invM = InvM_real + qpidx*Nsize*Nsize;
-
-  mat = (double *)malloc(sizeof(double)*n2*n2);
-  work = (double *)malloc(sizeof(double)*n2*n2);
-
-  /* X_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k;
-    vec_k = vec + k*nsize;
-    for(l=k+1;l<n;l++) {
-      val = 0.0;
-      for(msi=0;msi<nsize;msi++) {
-        val += w[l*nsize + msi] * vec_k[msi];
-      }
-      mat_k[l] = val + vec_k[msa[l]];
-    }
-  }
-
-  /* Y_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k + n;
-    for(l=0;l<n;l++) {
-      mat_k[l] = w[k*nsize + msa[l]];
-    }
-  }
-
-  /* Z_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*(k+n) + n;
-    invM_k = invM + msa[k]*nsize;
-    for(l=k+1;l<n;l++) {
-      mat_k[l] = invM_k[msa[l]];
-    }
-  }
-
-  //#pragma loop noalias
-  for(k=0;k<n2;k++) {
-    //#pragma loop norecurrence
-    for(l=0;l<k;l++) {
-      mat[n2*k + l] = -mat[n2*l + k]; /* transpose */
-    }
-    mat[n2*k + k] = 0.0; /* diagonal elements */
-  }
-
-  /* calculate Pf M */
-  M_DSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, &info);
-
-  sgn = ( (n*(n-1)/2)%2==0 ) ? 1.0 : -1.0;
-
-  //free(matUV);
-  free(mat);
-  //free(invMat);
-  free(work);
-  return sgn * pfaff * PfM_real[qpidx];
-}
-
-/* Calculate new pfaffian with Backflow effects.
-   The ma-th electron with spin s hops from ra to rb */
-void UpdateMAll_BF_real(const int *icount, const int *msaTmp,
-                  double *pfMNew, const int *eleIdx,
-                  const int qpStart, const int qpEnd) {
-#pragma procedure serial
-  const int qpNum = qpEnd-qpStart;
-  int qpidx;
-  int globalQpidx;
-  //double complex *sltE;
-  //double complex *sltE_i;
-  int *msa;
-  int i;
-  //int *hop;
-  //double complex diff;
-
-  for(qpidx=0;qpidx<qpNum;qpidx++) {
-    globalQpidx = qpidx + qpStart;
-    //Store msa//
-    msa=(int *)malloc(sizeof(int)*icount[globalQpidx]);
-    for(i=0;i<icount[globalQpidx];i++){
-      msa[i] = msaTmp[i+globalQpidx*Nsize];
-    }
-
-    /* calculateNewPfM */
-    pfMNew[qpidx] = updateMAll_BF_real_child(qpidx,globalQpidx,icount[globalQpidx],msa,eleIdx);
-
-    free(msa);
-  }
-
-  return;
-}
-
-/* msa[k]-th electron hops from rsa[k] to eleIdx[msa[k]] */
-/* buffer size = n*Nsize */
-//void updateMAllBF3_child(const int qpidx, const int n, const int *msa,
-double updateMAll_BF_real_child(const int qpidx, const int globalQpidx, const int n, const int *msa,
-                          const int *eleIdx) {
-  const int nsize = Nsize;
-  const int n2 = 2*n;
-  const double *sltE;
-  const double *sltE_k;
-  double *invM;
-  double *invM_i, *invM_k;
-
-  double *vec; /* vec[n][nsize] */
-  double *vec_k;
-  double *w; /* w[n][nsize] = vec * invM^T */
-  //double complex mat[n2*n2]; /* mat[n2][n2] */
-  double *mat; /* mat[n2][n2] */
-  double *mat_k;
-  //double complex invMat[n2*n2]; /* mat[n2][n2] */
-  double *invMat; /* mat[n2][n2] */
-  //double complex matUV[n2*nsize]; /* mat[n2][nsize] */
-  double *matUV; /* mat[n2][nsize] */
-  double *woodburyTmp; /* [nsize][n2] */
-  double *matUV_i;
-  double sgn;
-
-  int rsi,rsk,msi,msj,k,l;
-  double val;
-
-  /* for DSKPFA */
-  char uplo='U', mthd='P';
-  int m,nn,lda,info=0;
-  double pfaff;
-  int iwork[n2];
-  int iwork2[n2];
-  //double complex work[n2*n2]; /* [n2][n2] */
-  double *work; /* [n2][n2] */
-  double *work2; /* [n2][n2] */
-  int lwork = n2*n2;
-  int lwork2 = n2;
-  m=nn=lda=n2;
-
-  sltE = SlaterElmBF_real + globalQpidx*Nsite2*Nsite2;
-  invM = InvM_real + qpidx*Nsize*Nsize;
-
-  //vec = bufferc; /* n*nsize */
-  vec = (double *)malloc(sizeof(double)*n*nsize);
-  w = (double *)malloc(sizeof(double)*n*nsize);
-  invMat = (double *)malloc(sizeof(double)*n2*n2);
-  mat = (double *)malloc(sizeof(double)*n2*n2);
-  matUV = (double *)malloc(sizeof(double)*n2*nsize);
-  woodburyTmp = (double *)malloc(sizeof(double)*nsize*n2);
-  work = (double *)malloc(sizeof(double)*n2*n2);
-  work2 = (double *)malloc(sizeof(double)*n2);
-
-#pragma loop noalias
-  for(k=0;k<n;k++) {
-    rsk = eleIdx[msa[k]] + (msa[k]/Ne)*Nsite;
-    sltE_k = sltE + rsk*Nsite2;
-    vec_k = vec + k*nsize;
-#pragma loop norecurrence
-    for(msi=0;msi<nsize;msi++) {
-      rsi = eleIdx[msi] + (msi/Ne)*Nsite;
-      vec_k[msi] = sltE_k[rsi];
-    }
-  }
-
-  DgemmRowMajorNT(n, nsize, nsize, vec, invM, 1.0, 0.0, w);
-
-  /* X_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k;
-    vec_k = vec + k*nsize;
-    for(l=k+1;l<n;l++) {
-      val = 0.0;
-      for(msi=0;msi<nsize;msi++) {
-        val += w[l*nsize + msi] * vec_k[msi];
-      }
-      mat_k[l] = val + vec_k[msa[l]];
-    }
-  }
-
-  /* Y_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*k + n;
-    for(l=0;l<n;l++) {
-      mat_k[l] = w[k*nsize + msa[l]];
-    }
-  }
-
-  /* Z_kl */
-  for(k=0;k<n;k++) {
-    mat_k = mat + n2*(k+n) + n;
-    invM_k = invM + msa[k]*nsize;
-    for(l=k+1;l<n;l++) {
-      mat_k[l] = invM_k[msa[l]];
-    }
-  }
-
-#pragma loop noalias
-  for(k=0;k<n2;k++) {
-#pragma loop norecurrence
-    for(l=0;l<k;l++) {
-      mat[n2*k + l] = -mat[n2*l + k]; /* transpose */
-    }
-    mat[n2*k + k] = 0.0; /* diagonal elements */
-  }
-
-  //#pragma loop noalias
-  for(k=0;k<n2;k++) {
-    //#pragma loop norecurrence
-    for(l=0;l<n2;l++) {
-      invMat[n2*k + l] = mat[n2*l + k];
-    }
-  }
-
-  /* matUV_ik */
-  for(msi=0;msi<nsize;msi++) {
-    matUV_i = matUV + n2*msi;
-    invM_i = invM + msi*nsize;
-    for(k=0;k<n;k++) {
-      val = -w[k*nsize + msi];
-      if(msa[k]==msi){val -= 1.0;}
-      matUV_i[k] = val;
-      matUV_i[k+n] = invM_i[msa[k]];
-    }
-  }
-  /* DInv */
-  // NOTE: This routines is not called in current version (i.e. lcoal coverage 0),
-  //       hence cannot be tested. I'm keeping GETRF and GETRI here.
-  // TODO: If anyone's interested in utilizing this routine, she or he might want to
-  //       utilize DSKTRF+LTL2INV
-  m=nn=lda=n2;
-  //M_ZGETRF(&m, &nn, invMat, &lda, iwork2, &info); /* ipiv = iwork */
-  M_DGETRF(&m, &nn, invMat, &lda, iwork2, &info); /* ipiv = iwork */
-  if(info!=0){
-    printf("Error M_ZGETRF %d\n",info);
-    //return;
-  }
-
-  //M_ZGETRI(&nn, invMat, &lda, iwork2, work2, &lwork2, &info);
-  M_DGETRI(&nn, invMat, &lda, iwork2, work2, &lwork2, &info);
-  if(info!=0){
-    printf("Error M_ZGETRI %d\n",info);
-    //return;
-  }
-
-  DgemmRowMajorNN(nsize, n2, n2, matUV, invMat, 1.0, 0.0, woodburyTmp);
-  DgemmRowMajorNT(nsize, nsize, n2, woodburyTmp, matUV, -1.0, 1.0, invM);
-
-  for(msi=0;msi<Ne;msi++) {
-    //invM_i = invM + msi*nsize;
-    for(msj=Ne;msj<Nsize;msj++) {
-      invM[msj*nsize+msi] = -invM[msi*nsize+msj];
-    }
-  }
-  for(msi=Ne;msi<Nsize;msi++) {
-    for(msj=Ne;msj<Nsize;msj++) {
-      invM[msi*nsize+msj] = 0.0;
-    }
-  }
-  for(msi=0;msi<Ne;msi++) {
-    for(msj=0;msj<Ne;msj++) {
-      invM[msi*nsize+msj] = 0.0;
-    }
-  }
-
-  /* calculate Pf M */
-  //M_ZSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, rwork, &info);
-  //M_DSKPFA(&uplo, &mthd, &nn, invM, &lda, &pfaff, iwork, work, &lwork, &info);
-  M_DSKPFA(&uplo, &mthd, &nn, mat, &lda, &pfaff, iwork, work, &lwork, &info);
-
-  sgn = ( (n*(n-1)/2)%2==0 ) ? 1.0 : -1.0;
-
-  free(matUV);
-  free(mat);
-  free(w);
-  free(vec);
-  free(invMat);
-  free(woodburyTmp);
-  free(work);
-  free(work2);
-
-  return sgn * pfaff * PfM_real[qpidx];
-}
